+++ pype9/cmd/convert.py
@@ -3,7 +3,7 @@ Tool to convert 9ML files between different supported formats (e.g. XML_,
 JSON_, YAML_) and 9ML versions.
 """
 from argparse import ArgumentParser
-from ._utils import nineml_document
+from pype9.utils.arguments import nineml_document
 from pype9.utils.logging import logger
